pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voxmonitor"
version = "0.2.0"
description = "Pig vocalization classification and synthesis pipeline using DeepSuite"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Anton Feldmann" }]
keywords = ["audio", "classification", "pig-vocalization", "multitask-learning"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
  "deepsuite>=1.0.8",
  "torch>=2.6.0",
  "torchvision>=0.18.0",
  "pytorch-lightning>=2.6.0",
  "torchaudio>=2.6.0",
  "numpy>=1.24.0",
  "scipy>=1.10.0",
  "pandas>=2.0.0",
  "pyyaml>=6.0.0",
  "omegaconf>=2.3.0",
  "requests>=2.31.0",
  "loguru>=0.7.3",
  "soundfile>=0.13.0",
  "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.1.0",
  "mypy>=1.5.0",
  "pytest>=7.4.0",
  "black>=25.1.0",
]

[project.scripts]
voxmonitor-train = "voxmonitor.train:main"

[tool.hatch.build.targets.wheel]
packages = ["src/voxmonitor"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
